Remove support for old K&R style
Rationale:
The goal is to compile on modern compiler/architecture.
Supporting legacy style has a cost: even if we don't change code, we need to read code, review code, etc...
Anyone targetting a legacy architecture and compiler will just use netlib fdlibm.
Supporting both legacy and modern is not a goal of this project, so YAGNI.