Dead store
i0 and i1 are assigned but never used.
Note that the removed sentence `i0 = ((*(int*)&one)>>29)^1; i1=1-i0;` is a strange way to test for endianness.
i0 and i1 are then the offset to apply on (unsigned *) pointer aliasing to get high and low 32-bits respectively.
There are other instances of this heresy scattered in code.
One should better use __HI and __LO macros as a replacement.