Remove warning about uninitialized variable t.
The indentation is misleading.
The else branch on line 79 stops just after initializing variable `t`.
Thus `t` is not initialized in the if branch, but used afterwards.

This is a non problem, because the if branch always return control to upper level.
This is because `if(huge+x>one)` is always true.
The only purpose of such if is to raise inexact floating point status flag.

We could enclose all the indented instructions into braces { ... } in the else branch...
Or just remove the else, to let analyzers be sure that the  `else if (ix<0x3fe00000) {     /* |x|<0.5 */` branch always return.