make __incExp macro robust in case of negative increment
Exponent increment is no always positive.
It's the case at least in e_hypot.c where a scaling of +600 or -600 can be applied in order to avoid overflow/underflow.
Note that previous usage of <<20 on a negative signed integer was problematic as it is considered as undefined behavior.

TODO: analyze every remaining usage of left shifting on signed integer, particularly all the <<20 bit shifting for adjusting exponent.