Don't shift left negative integers.
For 2-complement, there is no problem related to shifting negative int, as long as the shift does not overflow the sign bit.
This is exactly the same than left shifting positive signed int: an overflow could as well overwrite sign bit and garbage the result.

But gurus of the standard decided it should be UB, despite a form of symmetry with positive case, and despite some programs like fdlibm did make a correct use of it (implementation defined due to 2 complement assumption, but then what? We don't need to target exotic micro controllers).
Let's abide to the new rules...